backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
gunicorn==22.0.0
redis==5.0.6
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pydantic[email]==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
alembic==1.13.1
python-multipart==0.0.9
Pillow==10.4.0
nudenet==3.4.2
httpx==0.27.2
sentry-sdk==2.19.2
